contrib/api_lead/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User

from .models import Item



class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'is_staff']


class ItemSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Item
        fields = ['url', 'name', 'description', 'price']
