requirements.txt
autopep8==1.4.4
backcall==0.1.0
captcha==0.3
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.13
django-allauth==0.40.0
django-appconf==1.0.3
django-axes==5.0.19
django-classy-tags==0.9.0
django-cms==3.6.0
django-filer==1.5.0
django-formtools==2.1
django-ipware==2.1.0
django-js-asset==1.2.2
django-location-field==2.1.0
django-map-widgets==0.2.2
django-mptt==0.10.0
django-osm-field==0.3.1
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-installer==1.1.1
djangocms-link==2.5.0
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
easy-thumbnails==2.6
html5lib==1.0.1
idna==2.8
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
traitlets==4.3.3
tzlocal==2.0.0
Unidecode==1.0.23
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
